requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
aiohttp
azure-eventhub>=5.11.2
openai>=1.0.0
azure-eventhub
azurefunctions-extensions-http-fastapi
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.14.1
httpx>=0.24.0
azure-cosmos>=4.5.1
pytest>=7.4.0